Move ManifestInfo serialization to separate namespace
