+++ src/TemplateGenerator/TemplateGeneratorCli.cs
@@ -3,9 +3,11 @@
 
 using ConsoleAppFramework;
 using Microsoft.DotNet.ImageBuilder.ReadModel;
+using Microsoft.DotNet.ImageBuilder.ReadModel.Serialization;
 using Microsoft.DotNet.DockerTools.Templating.Abstractions;
 using Cottle;
 using Microsoft.DotNet.DockerTools.Templating.Cottle;
+using Microsoft.DotNet.ImageBuilder.Models.Manifest;
 
 namespace Microsoft.DotNet.DockerTools.TemplateGenerator;
 
@@ -19,8 +21,6 @@ public sealed class TemplateGeneratorCli
     public async Task GenerateDockerfiles([Argument] string manifestPath)
     {
         ManifestInfo manifest = await ManifestInfo.LoadAsync(manifestPath);
-        var manifestString = manifest.ToJsonString();
-        await File.WriteAllTextAsync("manifest.processed.json", manifestString);
     }
 
     [Command("generate-template")]
